+++ codepropertygraph/src/main/java/io/shiftleft/codepropertygraph/cpgloading/Defaults.java
@@ -1,4 +1,4 @@
-package io.shiftleft.cpgloading;
+package io.shiftleft.codepropertygraph.cpgloading;
 
 public class Defaults {
