+++ codepropertygraph/src/main/scala/io/shiftleft/codepropertygraph/cpgloading/IgnoredProtoEntries.scala
@@ -1,4 +1,4 @@
-package io.shiftleft.cpgloading
+package io.shiftleft.codepropertygraph.cpgloading
 
 /* some non-public frontends (e.g. java2cpg) use cpg proto entries that are `UNRECOGNIZED` by cpg-public.
  we'll ignore those during the import */
