+++ codepropertygraph/src/main/scala/io/shiftleft/codepropertygraph/cpgloading/OnDiskOverflowConfig.scala
@@ -1,6 +1,6 @@
-package io.shiftleft.cpgloading
+package io.shiftleft.codepropertygraph.cpgloading
 
-import io.shiftleft.cpgloading.Defaults.DEFAULT_CACHE_HEAP_PERCENTAGE
+import io.shiftleft.codepropertygraph.cpgloading.Defaults.DEFAULT_CACHE_HEAP_PERCENTAGE
 
 /** configure graphdb to use ondisk overflow. by default, system tmp directory is used (e.g. `/tmp`)  */
 case class OnDiskOverflowConfig(cacheHeapPercentage: Float = DEFAULT_CACHE_HEAP_PERCENTAGE,
