+++ cpgqueryingtests/src/main/scala/io/shiftleft/cpgqueryingtests/codepropertygraph/CpgFactory.scala
@@ -3,16 +3,14 @@ package io.shiftleft.cpgqueryingtests.codepropertygraph
 import java.io.{File, PrintWriter}
 import java.nio.file.Files
 
+import io.shiftleft.codepropertygraph.Cpg
 import io.shiftleft.cpgloading.{CpgLoader, CpgLoaderConfig}
-import io.shiftleft.queryprimitives.steps.starters.Cpg
 
-class CpgFactory(frontend: LanguageFrontend,
-                 semanticsFilename : String) {
+class CpgFactory(frontend: LanguageFrontend, semanticsFilename: String) {
 
   /**
     * Build a CPG for the provided C/C++ code snippet.
     * */
-
   def buildCpg(sourceCode: String): Cpg = {
     val tmpDir = Files.createTempDirectory("dflowtest").toFile
     tmpDir.deleteOnExit()
