+++ cpgqueryingtests/src/test/scala/io/shiftleft/cpgqueryingtests/steps/CpgDataFlowTests.scala
@@ -1,13 +1,12 @@
 package io.shiftleft.cpgqueryingtests.steps
 
+import io.shiftleft.codepropertygraph.Cpg
 import io.shiftleft.codepropertygraph.generated.nodes
-import io.shiftleft.queryprimitives.steps.starters.Cpg
 import io.shiftleft.queryprimitives.steps._
 import io.shiftleft.queryprimitives.steps.types.expressions.Literal
 import io.shiftleft.queryprimitives.steps.types.structure.{Member, Method}
 import io.shiftleft.queryprimitives.steps.ext.dataflowengine._
 import org.scalatest.{Matchers, WordSpec}
-
 import shapeless.HNil
 
 class CpgDataFlowTests extends WordSpec with Matchers {
