+++ query-primitives/build.sbt
@@ -1,10 +1,9 @@
 name := "query-primitives"
 
-dependsOn(Projects.codepropertygraph, Projects.protoBindings)
+dependsOn(Projects.codepropertygraph)
 
 libraryDependencies ++= Seq(
   "org.json4s" %% "json4s-native" % "3.5.2",
-  "commons-io" % "commons-io" % "2.5",
   "org.scala-lang.modules" %% "scala-java8-compat" % "0.8.0",
   "com.massisframework" % "j-text-utils" % "0.3.4",
   "org.scalatest" %% "scalatest" % "3.0.3" % Test
