+++ query-primitives/src/main/scala/io/shiftleft/queryprimitives/steps/ext/queryprimitives/package.scala
@@ -1,0 +1,10 @@
+package io.shiftleft.queryprimitives.steps.ext
+
+import io.shiftleft.codepropertygraph.Cpg
+import io.shiftleft.queryprimitives.steps.starters.NodeTypeStarters
+
+package object queryprimitives {
+
+  implicit def toNodeTypeStarters(cpg: Cpg): NodeTypeStarters =
+    new NodeTypeStarters(cpg)
+}
